requirements-dev.txt

-r requirements.txt
black
isort
pre-commit
pylance
pylint[spelling]
pytest
pytest-flask
